add bs.mksh
this bs.mksh works differently. it assumes that you
have a as an array, l as the lowest index value to search from,
h as the highest index to search to and i as the item to search for.

after setting these, source the file.
if m is not -1, then the item is found at index $m